rename the gem name
